# server/openapi-definition/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "BMS API Endpoints",
    "description" : "Here is description of BMS API.",
    "contact" : {
      "name" : "BMS administrator",
      "email" : "contact@email.example.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8888",
    "description" : "Local server"
  }, {
    "url" : "http://192.168.0.119:8888",
    "description" : "A remote server"
  } ],
  "tags" : [ {
    "name" : "BMS basic API",
    "description" : "BMS Common APIs for authenticated users"
  }, {
    "name" : "BMS guest API",
    "description" : "API endpoints for guest users"
  } ],
  "paths" : {
    "/api/v1/guest/players" : {
      "get" : {
        "tags" : [ "BMS guest API" ],
        "parameters" : [ {
          "name" : "team_ids",
          "in" : "query",
          "description" : "Filter with team ids",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "int32",
              "type" : "integer"
            }
          },
          "style" : "form",
          "explode" : false
        }, {
          "name" : "X-Foo",
          "in" : "header",
          "description" : "test header parameter",
          "schema" : {
            "default" : "default-foo-value",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a list of players",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guest/teams" : {
      "get" : {
        "tags" : [ "BMS guest API" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Team"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players" : {
      "get" : {
        "tags" : [ "BMS basic API" ],
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "default" : 50,
            "type" : "number"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "team_ids",
          "in" : "query",
          "description" : "filter with team ids",
          "schema" : {
            "type" : "array",
            "items" : {
              "format" : "int32",
              "type" : "integer"
            }
          },
          "style" : "form",
          "explode" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a list of players",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams" : {
      "get" : {
        "tags" : [ "BMS basic API" ],
        "summary" : "Return a list of teams",
        "description" : "Return a list of teams",
        "responses" : {
          "200" : {
            "description" : "Returns a list of teams",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Team"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        }, {
          "jwt token" : [ ]
        } ]
      }
    },
    "/api/v1/teams/{id}" : {
      "get" : {
        "tags" : [ "BMS basic API" ],
        "summary" : "Return a team detail",
        "description" : "Return a team detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the detail of a team",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "name" : {
                      "type" : "string"
                    },
                    "url_path" : {
                      "maxLength" : 32,
                      "type" : "string"
                    },
                    "at_bats" : {
                      "format" : "double",
                      "maximum" : 5,
                      "minimum" : 0,
                      "type" : "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{id}/players" : {
      "get" : {
        "tags" : [ "BMS basic API" ],
        "summary" : "Return players of the team",
        "description" : "Return players of the team",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return players of the team",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Player" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "team" : {
            "$ref" : "#/components/schemas/Team"
          },
          "backNumber" : {
            "type" : "string"
          }
        }
      },
      "Team" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "url_path" : {
            "maxLength" : 32,
            "type" : "string"
          },
          "at_bats" : {
            "format" : "double",
            "maximum" : 5,
            "minimum" : 0,
            "type" : "number"
          }
        }
      }
    },
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "jwt token" : {
        "type" : "apiKey",
        "in" : "header",
        "scheme" : "bearer",
        "bearerFormat" : "jwt"
      }
    }
  }
}